# mxsuite-api/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: MXSuite API
  description: |
    MXSuite Data Onboarding Platform — REST API.

    ## Authentication
    All endpoints (except `/api/auth/login` and `/actuator/health`) require a Bearer JWT token
    in the `Authorization` header.

    Platform users (`PLATFORM_ADMIN`, `PLATFORM_SUPPORT`) may include an `X-Tenant-Id` header
    to operate within a specific customer tenant's context.

    ## Pagination
    List endpoints accept Spring Data `Pageable` query parameters:
    - `page` — zero-based page index (default `0`)
    - `size` — page size (default `20`)
    - `sort` — property,direction (e.g. `createdAt,desc`)

    ## Multi-Tenancy
    All tenant-scoped resources are automatically filtered by the tenant context
    derived from the authenticated user's JWT token.
  version: 0.1.0
  contact:
    name: MXSuite Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects for current tenant
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProject'
              example:
                content:
                  - id: p1111111-1111-1111-1111-111111111111
                    name: Member Import 2026
                    description: Annual member data import from legacy CRM
                    createdAt: "2026-07-10T09:00:00Z"
                    lastModifiedAt: "2026-07-14T16:30:00Z"
                    owner:
                      id: b2c3d4e5-f6a7-8901-bcde-f12345678901
                      firstName: Denver
                      lastName: Admin
                totalElements: 1
                totalPages: 1
                size: 20
                number: 0

    post:
      tags: [Projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: Member Import 2026
              description: Annual member data import from legacy CRM
      responses:
        '200':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: p1111111-1111-1111-1111-111111111111
                name: Member Import 2026
                description: Annual member data import from legacy CRM
                createdBy: b2c3d4e5-f6a7-8901-bcde-f12345678901
                createdAt: "2026-07-15T14:30:00Z"
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get project by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: p1111111-1111-1111-1111-111111111111
                name: Member Import 2026
                description: Annual member data import from legacy CRM
                createdAt: "2026-07-10T09:00:00Z"
                lastModifiedAt: "2026-07-14T16:30:00Z"
        '404':
          description: Project not found

    put:
      tags: [Projects]
      summary: Update a project
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: Member Import 2026 — Revised
              description: Updated scope to include event data
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: p1111111-1111-1111-1111-111111111111
                name: Member Import 2026 — Revised
                description: Updated scope to include event data
                lastModifiedAt: "2026-07-15T16:00:00Z"
        '404':
          description: Project not found

    delete:
      tags: [Projects]
      summary: Delete a project
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Project deleted
        '404':
          description: Project not found

  # ──────────────────────────────────────────────
  # Plans
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /api/auth/login
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: "22222222-2222-2222-2222-222222222222"
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Zero-based page index
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Page size
    sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: "Sort property,direction (e.g. createdAt,desc)"
      example: createdAt,desc
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        tenantType:
          $ref: '#/components/schemas/TenantType'
        active:
          type: boolean
        createdBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastModifiedBy:
          type: string
          format: uuid
          nullable: true
        lastModifiedAt:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
          description: Job title
        bio:
          type: string
          nullable: true
          description: Short biography (max 500 chars)
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        preferences:
          type: object
          nullable: true
          additionalProperties: true
        tenant:
          $ref: '#/components/schemas/Tenant'
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        owner:
          $ref: '#/components/schemas/User'
        tenant:
          $ref: '#/components/schemas/Tenant'
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true

    # ── Plan ──
...
    TenantType:
      type: string
      enum: [PLATFORM, CUSTOMER]
    UserRole:
      type: string
      enum: [PLATFORM_ADMIN, PLATFORM_SUPPORT, TENANT_ADMIN, TENANT_USER]
...
    PageProject:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...